requirements.txt
# ========================================
# TalentScout - Complete Python Requirements
# All-in-One: Backend + Streamlit UI
# ========================================

# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9

# AI & LLM - Google Gemini 2.0
google-generativeai==0.3.2
chromadb==0.4.22

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# HTTP & API
httpx==0.26.0
requests==2.31.0

# Data Processing
pandas==2.1.4
numpy==1.26.3

# Validation
email-validator==2.1.0
phonenumbers==8.13.27

# Logging
python-json-logger==2.0.7

# Streamlit UI
streamlit==1.29.0
watchdog==3.0.0
plotly==5.18.0

# Development & Testing
pytest==7.4.4
pytest-asyncio==0.23.3

# Date & Time
python-dateutil==2.8.2

# CORS
python-cors==1.0.0
